Make event loop default for httpRemoteTask related code
