+++ presto-main-base/src/main/java/com/facebook/presto/execution/TaskManagerConfig.java
@@ -99,7 +99,6 @@ public class TaskManagerConfig
     private double highMemoryTaskKillerGCReclaimMemoryThreshold = 0.01;
     private Duration highMemoryTaskKillerFrequentFullGCDurationThreshold = new Duration(1, SECONDS);
     private double highMemoryTaskKillerHeapMemoryThreshold = 0.9;
-    private boolean enableEventLoop = true;
     private Duration slowMethodThresholdOnEventLoop = new Duration(0, SECONDS);
 
     public long getSlowMethodThresholdOnEventLoop()
@@ -112,18 +111,6 @@ public class TaskManagerConfig
     {
         this.slowMethodThresholdOnEventLoop = slowMethodThresholdOnEventLoop;
         return this;
-    }
-
-    @Config("task.enable-event-loop")
-    public TaskManagerConfig setEventLoopEnabled(boolean enableEventLoop)
-    {
-        this.enableEventLoop = enableEventLoop;
-        return this;
-    }
-
-    public boolean isEventLoopEnabled()
-    {
-        return enableEventLoop;
     }
 
     @MinDuration("1ms")
